Big update
Finally able to live-stream from Raspberry Pi camera to the central server.
The transmission is a little bit laggy, so there's room for improvement.
But overall, it's comprehensive.

+++ hwsystem/HWSystem.py
@@ -1,4 +1,4 @@
-import cv2
+import cv2 as cv
 import os
 import requests
 import json
@@ -32,7 +32,14 @@ class HWSystem:
         self.registration_trials = 0
         self.register_json_path = register_json_path
         self.config_json_path = config_json_path
-
+        
+        self.RASPICAM_ID = 0
+        self.OUTPUT_CHUNK_NAME = 'framechunk'
+        
+        
+        self.video_capturer = None
+        self.video_writer = None
+        
         with open(self.register_json_path) as register_data:
             self.REGISTER_JSON = json.load(register_data)
         self.REGISTER_JSON['access_key'] = self.ACCESS_KEY
@@ -158,6 +165,7 @@ class HWSystem:
         Stops the livestream.
         """
         self.stop_livestream.set()
+        self.video_capturer.release()
         print('[INFO] Stopped livestreaming!')
 
     def listen_for_sse(self) -> None:
@@ -268,27 +276,78 @@ class HWSystem:
 
         print('[SUCCESS] Renewed JWT Token!')
 
+    def record_chunk(self) -> None:
+        i = 0
+        
+        output_file_name = self.OUTPUT_CHUNK_NAME + '.' + self.CONFIG_JSON['Preprocess']['DataFormat']
+        FRAME_WIDTH = int(self.video_capturer.get(cv.CAP_PROP_FRAME_WIDTH))
+        FRAME_HEIGHT = int(self.video_capturer.get(cv.CAP_PROP_FRAME_HEIGHT))
+        
+        # VideoWriter object to save the recorded video.
+        self.video_writer = cv.VideoWriter(output_file_name,
+                             cv.VideoWriter_fourcc(*'mp4v'), 
+                             self.CONFIG_JSON['Preprocess']['ChunkSize'],
+                             (FRAME_WIDTH, FRAME_HEIGHT))
+                             
+        while True:
+            # Read a frame from the camera.
+            ret, frame = self.video_capturer.read()
+            
+            # Check if the frame has been successfully read.
+            if not ret:
+                print('[ERROR] Failed to read the frame!')
+                print('Aborting...')
+                exit(1)
+                
+            # Write the frame to the output video.
+            self.video_writer.write(frame) 
+            
+            # Exit the loop if 'q' is pressed.
+            if cv.waitKey(1) & 0xFF == ord('q'):
+                break
+                
+            i += 1
+            if i >= self.CONFIG_JSON['Preprocess']['ChunkSize']:
+                break
+                
+        self.video_writer.release()
+        
+        
     def livestream(self) -> None:
         """
         Perform the livestreaming to the central server.
         """
+        output_file_name = self.OUTPUT_CHUNK_NAME + '.' + self.CONFIG_JSON['Preprocess']['DataFormat']
+        
         while not self.stop_livestream.is_set():
             print('[INFO] Livestreaming to the central server...')
-            with open("actual_violence.mp4", 'rb') as f:
+            self.record_chunk()
+            
+            with open(output_file_name, 'rb') as f:
+                print(self.BASE_ROUTE)
                 response = requests.post(f'{self.BASE_ROUTE}/api/send/frames',
-                                         files={"Video2.mp4": f},
+                                         files={output_file_name: f},
                                          headers={'Authorization': 'Bearer ' + self.CONFIG_JSON['JWT Token']})
                 response_json = response.json()
-
-                if response_json['Code'] == 400:
-                    self.handle_sse_stop()
-                    return
-                elif response_json['Code'] == 403:
-                    if os.path.exists(self.config_json_path):
-                       os.remove(self.config_json_path) 
-    
-                self.renew_jwt_token(response.json())
-            sleep(5)
+            
+            if 'Code' not in response_json:
+                self.video_capturer.release()
+                raise requests.exceptions.ConnectionError()
+            
+            if response_json['Code'] == 400:
+                self.handle_sse_stop()
+                return
+            elif response_json['Code'] == 403:
+                if os.path.exists(self.config_json_path):
+                    os.remove(self.config_json_path)
+            
+            # Write the frames again from the beginning.
+            #self.video_writer.set(cv.CAP_PROP_POS_FRAMES, 0)
+            
+            # Renewing the JWT Token
+            self.renew_jwt_token(response.json())
+            
+        cap.release()
 
     def start_module(self) -> None:
         """
@@ -297,7 +356,17 @@ class HWSystem:
         while True:
             try:
                 print("[START] Raspberry PI module...")
+                
+                # Open the video capture.
+                self.video_capturer = cv.VideoCapture(self.RASPICAM_ID)
 
+                # Check if the camera has ben successfully opened.
+                if not self.video_capturer.isOpened():
+                        print('[ERROR] Failed to open the camera!')
+                        print('Aborting...')
+                        exit(1)
+                
+        
                 register_is_success = None
                 if not os.path.exists(self.config_json_path):
                     register_is_success = self.attempt_registration()
